# Optimize geographical area association