--- a/app/controllers/api/v1/geographical_areas_controller.rb
+++ b/app/controllers/api/v1/geographical_areas_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V1
     class GeographicalAreasController < ApiController
       def countries
-        @geographical_areas = GeographicalArea.eager(:geographical_area_descriptions)
+        @geographical_areas = GeographicalArea.eager_graph(:geographical_area_descriptions)
                                               .actual
                                               .countries
                                               .all