This PR fixes a bug with sequel associations not caching or caching when is not required.

This will speed up the response times:


Before:

```
[2016-07-28T23:44:43.204807 #1652]  INFO -- : Started GET "/commodities/6202129090?as_of=2016-07-28" for 127.0.0.1 at 2016-07-28 23:44:43 -0500
[2016-07-28T23:44:43.266070 #1652]  INFO -- : Processing by Api::V1::CommoditiesController#show as JSON
[2016-07-28T23:44:43.266242 #1652]  INFO -- :   Parameters: {"as_of"=>"2016-07-28", "id"=>"6202129090"}
[2016-07-28T23:44:51.072439 #1652]  INFO -- :   Rendered api/v1/commodities/show.json.rabl (6648.0ms)
[2016-07-28T23:44:51.072976 #1652]  INFO -- : Completed 200 OK in 7807ms (Views: 3209.5ms | Models: 3906.9ms)
```

After:

```
[2016-07-28T23:42:21.006828 #1450]  INFO -- : Started GET "/commodities/6202129090?as_of=2016-07-28" for 127.0.0.1 at 2016-07-28 23:42:21 -0500
[2016-07-28T23:42:21.073762 #1450]  INFO -- : Processing by Api::V1::CommoditiesController#show as JSON
[2016-07-28T23:42:21.073942 #1450]  INFO -- :   Parameters: {"as_of"=>"2016-07-28", "id"=>"6202129090"}
[2016-07-28T23:42:23.506844 #1450]  INFO -- :   Rendered api/v1/commodities/show.json.rabl (1234.4ms)
[2016-07-28T23:42:23.507241 #1450]  INFO -- : Completed 200 OK in 2433ms (Views: 751.9ms | Models: 971.9ms)
```